# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ApplicationUserList": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/AuthenticatedUser"
            },
            "title": "Results",
            "type": "array"
          }
        },
        "required": [
          "count",
          "results"
        ],
        "title": "ApplicationUserList",
        "type": "object"
      },
      "ApplicationUserProfileUpdate": {
        "properties": {
          "displayName": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Displayname",
            "type": "string"
          }
        },
        "required": [
          "displayName"
        ],
        "title": "ApplicationUserProfileUpdate",
        "type": "object"
      },
      "ApplicationUserRoleUpdate": {
        "properties": {
          "role": {
            "enum": [
              "annotator",
              "reviewer",
              "admin"
            ],
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "role"
        ],
        "title": "ApplicationUserRoleUpdate",
        "type": "object"
      },
...
      "AuthenticatedUser": {
        "properties": {
          "disabled": {
            "default": false,
            "title": "Disabled",
            "type": "boolean"
          },
          "displayName": {
            "title": "Displayname",
            "type": "string"
          },
          "email": {
            "maxLength": 320,
            "minLength": 3,
            "title": "Email",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "role": {
            "enum": [
              "annotator",
              "reviewer",
              "admin"
            ],
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "displayName",
          "role"
        ],
        "title": "AuthenticatedUser",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PipelineEventDto": {
        "properties": {
          "createdAt": {
            "title": "Createdat",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "phase": {
            "title": "Phase",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "phase",
          "status",
          "message",
          "createdAt"
        ],
        "title": "PipelineEventDto",
        "type": "object"
      },
      "PipelineLogDto": {
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "timestamp": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp"
          }
        },
        "required": [
          "message"
        ],
        "title": "PipelineLogDto",
        "type": "object"
      },
      "PipelineProgressDto": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "percent": {
            "title": "Percent",
            "type": "integer"
          },
          "phase": {
            "title": "Phase",
            "type": "string"
          }
        },
        "required": [
          "phase",
          "percent",
          "detail"
        ],
        "title": "PipelineProgressDto",
        "type": "object"
      },
      "PipelineRunDto": {
        "properties": {
          "batchJobId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Batchjobid"
          },
          "canonicalPrefix": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Canonicalprefix"
          },
          "createdAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Createdat"
          },
          "datasetType": {
            "title": "Datasettype",
            "type": "string"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/PipelineEventDto"
            },
            "title": "Events",
            "type": "array"
          },
          "finishedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finishedat"
          },
          "images": {
            "default": 0,
            "title": "Images",
            "type": "integer"
          },
          "logs": {
            "items": {
              "$ref": "#/components/schemas/PipelineLogDto"
            },
            "title": "Logs",
            "type": "array"
          },
          "objects": {
            "default": 0,
            "title": "Objects",
            "type": "integer"
          },
          "release": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Release"
          },
          "requestGcsUri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requestgcsuri"
          },
          "requestedBy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requestedby"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "runId": {
            "title": "Runid",
            "type": "string"
          },
          "split": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Split"
          },
          "stages": {
            "items": {
              "$ref": "#/components/schemas/PipelineProgressDto"
            },
            "title": "Stages",
            "type": "array"
          },
          "startedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Startedat"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "validation": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validation"
          }
        },
        "required": [
          "runId",
          "datasetType",
          "status"
        ],
        "title": "PipelineRunDto",
        "type": "object"
      },
      "PipelineRunListDto": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/PipelineRunDto"
            },
            "title": "Results",
            "type": "array"
          }
        },
        "required": [
          "count",
          "results"
        ],
        "title": "PipelineRunListDto",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "QA orchestration API for 2D perception annotations",
    "title": "Label Guardian API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/auth/me": {
      "get": {
        "operationId": "get_my_profile_api_v1_auth_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticatedUser"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get My Profile",
        "tags": [
          "Authentication"
        ]
      },
      "put": {
        "operationId": "update_my_profile_api_v1_auth_me_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicationUserProfileUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticatedUser"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update My Profile",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/users": {
      "get": {
        "operationId": "list_application_users_api_v1_auth_users_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationUserList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Application Users",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/users/{user_id}/role": {
      "patch": {
        "operationId": "update_application_user_role_api_v1_auth_users__user_id__role_patch",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicationUserRoleUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticatedUser"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Application User Role",
        "tags": [
          "Authentication"
        ]
      }
    },
...
    "/api/v1/ingestion/runs": {
      "get": {
        "operationId": "list_runs_api_v1_ingestion_runs_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineRunListDto"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Runs",
        "tags": [
          "Cloud Ingestion"
        ]
      }
    },
    "/api/v1/ingestion/runs/{run_id}": {
      "get": {
        "operationId": "get_run_api_v1_ingestion_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineRunDto"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Run",
        "tags": [
          "Cloud Ingestion"
        ]
      }
    },
...
  }
}